backend/requirements.txt
fastapi
uvicorn
motor
pydantic
pydantic-settings
python-dotenv
openai
google-generativeai
httpx
jinja2
